function testOutput:
- special treatment of runtime limit parameter when its value equals zero
- possibility to leave running process (runtime limit parameter equals zero) and to call testOutput more times to assert various reaction during running

Added new redis test.

Added new Simple-agent test (it asserts reactions to the redis disconnection and reconnection)

+++ scripts/test_examples.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 
 TMP_FILE="/tmp/out"
+TMP_FILE2="/tmp/unprocessed"
 exitCode=0
 PREV_IFS="$IFS"
 
@@ -9,18 +10,31 @@ PREV_IFS="$IFS"
 # 1-st command to run
 # 2-nd array of expected strings in the command output
 # 3-rd argument is an optional array of unexpected strings in the command output
-# 4-th argument is an optional command runtime limit
+# 4-th argument is an optional command runtime limit (value 0 does not cause stopping of the run.)
 function testOutput {
 IFS="${PREV_IFS}"
 
     #run the command
     if [ $# -ge 4 ]; then
-        $1 > ${TMP_FILE} 2>&1 &
-        CMD_PID=$!
-        sleep $4
-        kill $CMD_PID
+        if [ -e ${TMP_FILE} ]; then
+	    # if exists file /tmp/out we assume that the command still runs, do not start it again
+            echo > /dev/null
+        else
+            $1 > ${TMP_FILE} 2>&1 &
+            CMD_PID=$!
+        fi
+        # do we set runtime limit?
+        if [ $4 -ne 0 ]; then
+            sleep $4
+            kill $CMD_PID
+        else
+	    # The command continues to run - his PID is in the variable CMD_PID
+            sleep 20
+        fi
+	cat ${TMP_FILE} | awk "NR > $processedLines" > ${TMP_FILE2}
     else
         $1 > ${TMP_FILE} 2>&1
+	cat ${TMP_FILE} > ${TMP_FILE2}
     fi
 
 IFS="
@@ -30,7 +44,7 @@ IFS="
     # loop through expected lines
     for i in $2
     do
-        if grep -- "${i}" /tmp/out > /dev/null ; then
+        if grep -- "${i}" ${TMP_FILE2} > /dev/null ; then
             echo "OK - '$i'"
         else
             echo "Not found - '$i'"
@@ -41,7 +55,7 @@ IFS="
     if [[ ! -z $3 ]] ; then
         for i in $3
         do
-            if grep -- "${i}" /tmp/out > /dev/null ; then
+            if grep -- "${i}" ${TMP_FILE2} > /dev/null ; then
                 echo "IS NOT OK - '$i'"
                 rv=1
             fi
@@ -50,12 +64,21 @@ IFS="
 
     # if an error occurred print the output
     if [[ ! $rv -eq 0 ]] ; then
-        cat ${TMP_FILE}
+        cat ${TMP_FILE2}
         exitCode=1
     fi
 
     echo "================================================================"
-    rm ${TMP_FILE}
+    if [ $4 -ne 0 ]; then
+        rm ${TMP_FILE}
+	rm ${TMP_FILE2}
+    else
+        # The command continues to run - the output is still redirected to the file ${TMP_FILE}
+	# read -n1 -r -p "Press any key to continue..." key
+	processedLines=`wc -l ${TMP_FILE} | cut --delimiter=" " -f1`
+	echo $processedLines
+	read -n1 -r -p "Press any key to continue..." key
+    fi
     return ${rv}
 }
 
@@ -109,6 +132,7 @@ function startCassandra {
     docker run -p 9042:9042 --name cassandra01 -d cassandra > /dev/null 2> /dev/null
     # Wait until cassandra is ready to accept a connection.
     for attemptps in {1..20} ; do
+        sleep 5
         NODEINFO=$(docker exec -it cassandra01 nodetool info)
         if [ $? -eq 0 ]; then
             if [[ ${NODEINFO} == *"Native Transport active: true"* ]]; then
@@ -123,6 +147,17 @@ function startCassandra {
 function stopCassandra {
     docker stop cassandra01 > /dev/null
     docker rm cassandra01 > /dev/null
+}
+
+function startRedis {
+    docker run --name redis -p 6379:6379 -d redis > /dev/null 2> /dev/null 
+    # sometimes Cassandra needs a bit more time to fully initialize
+    # sleep 2
+}
+
+function stopRedis {
+    docker stop redis > /dev/null
+    docker rm redis > /dev/null
 }
 
 #### Cassandra ###########################################################
@@ -140,6 +175,48 @@ cmd="examples/cassandra-lib/cassandra-lib examples/cassandra-lib/client-config.y
 testOutput "${cmd}" "${expected}"
 
 stopCassandra
+
+#### Redis ###########################################################
+
+startRedis
+
+expected=("config: redis.NodeConfig:
+INFO[0004] GetValue(key1) = true ; val = val 1 ; revision = 0  loc=\"simple/simple.go(208)\" logger=defaultLogger tag=00000000
+INFO[0004] GetValue(key2) = true ; val = val 2 ; revision = 0  loc=\"simple/simple.go(208)\" logger=defaultLogger tag=00000000
+==> NOTE: key3 should have expired
+INFO[0004] GetValue(key3) = false                        loc=\"simple/simple.go(210)\" logger=defaultLogger tag=00000000
+==> NOTE: get(key) should return false
+INFO[0004] GetValue(key) = false                         loc=\"simple/simple.go(210)\" logger=defaultLogger tag=00000000
+INFO[0004] ListKeys(key):  key2 (rev 0)                  loc=\"simple/simple.go(228)\" logger=defaultLogger tag=00000000
+INFO[0004] ListKeys(key):  key1 (rev 0)                  loc=\"simple/simple.go(228)\" logger=defaultLogger tag=00000000
+INFO[0004] ListKeys(key): count = 2                      loc=\"simple/simple.go(231)\" logger=defaultLogger tag=00000000
+INFO[0004] ListValues(key):  key2 = val 2 (rev 0)        loc=\"simple/simple.go(249)\" logger=defaultLogger tag=00000000
+INFO[0004] ListValues(key):  key1 = val 1 (rev 0)        loc=\"simple/simple.go(249)\" logger=defaultLogger tag=00000000
+INFO[0004] ListValues(key): count = 2                    loc=\"simple/simple.go(252)\" logger=defaultLogger tag=00000000
+INFO[0004] doKeyInterator(): Expected 100 keys; Found 100  loc=\"simple/simple.go(277)\" logger=defaultLogger tag=00000000
+INFO[0004] doKeyValInterator(): Expected 100 keyVals; Found 100  loc=\"simple/simple.go(315)\" logger=defaultLogger tag=00000000
+INFO[0004] Delete(key): found = true                     loc=\"simple/simple.go(341)\" logger=defaultLogger tag=00000000
+==> NOTE: All keys should have been deleted
+INFO[0004] GetValue(key1) = false                        loc=\"simple/simple.go(210)\" logger=defaultLogger tag=00000000
+INFO[0004] GetValue(key2) = false                        loc=\"simple/simple.go(210)\" logger=defaultLogger tag=00000000
+INFO[0004] ListKeys(key): count = 0                      loc=\"simple/simple.go(231)\" logger=defaultLogger tag=00000000
+INFO[0004] ListValues(key): count = 0                    loc=\"simple/simple.go(252)\" logger=defaultLogger tag=00000000
+INFO[0004] txn(): keys = [key101 key102 key103 key104]   loc=\"simple/simple.go(353)\" logger=defaultLogger tag=00000000
+INFO[0004] ListValues(key):  key102 = 2 (rev 0)          loc=\"simple/simple.go(249)\" logger=defaultLogger tag=00000000
+INFO[0004] ListValues(key):  key104 = 4 (rev 0)          loc=\"simple/simple.go(249)\" logger=defaultLogger tag=00000000
+INFO[0004] ListValues(key):  key103 = 3 (rev 0)          loc=\"simple/simple.go(249)\" logger=defaultLogger tag=00000000
+INFO[0004] ListValues(key): count = 3                    loc=\"simple/simple.go(252)\" logger=defaultLogger tag=00000000
+INFO[0004] Sleep for 5 seconds                           loc=\"simple/simple.go(175)\" logger=defaultLogger tag=00000000
+INFO[0009] Closing connection                            loc=\"simple/simple.go(179)\" logger=defaultLogger tag=00000000
+==> NOTE: Call on a closed connection should fail.
+ERRO[0009] Delete(key) called on a closed connection     loc=\"simple/simple.go(338)\" logger=defaultLogger tag=00000000
+INFO[0009] Sleep for 10 seconds                          loc=\"simple/simple.go(186)\" logger=defaultLogger tag=00000000
+")
+
+cmd="examples/redis-lib/simple/simple -redis-config=node-client.yaml "
+testOutput "${cmd}" "${expected}"
+
+stopRedis
 
 #### Configs #############################################################
 
@@ -515,6 +592,64 @@ testOutput "${cmd}" "${expected}" "${unexpected}" 5
 
 stopEtcd
 stopKafka
+
+#### Simple-agent with Cassandra and Redis and Kafka and ETCD ####################################
+
+startEtcd
+startCustomizedKafka examples/kafka-plugin/manual-partitioner/server.properties
+startRedis
+startCassandra
+
+expected=("Plugin etcdv3: status check probe registered
+Plugin redis: status check probe registered
+Plugin cassandra: status check probe registered
+Plugin kafka: status check probe registered
+All plugins initialized successfully
+Agent plugin state update.*plugin=etcdv3 state=ok
+Agent plugin state update.*plugin=redis state=ok
+Agent plugin state update.*plugin=cassandra state=ok
+Agent plugin state update.*plugin=kafka state=ok
+")
+
+unexpected=("redis config not found  - skip loading this plugin
+cassandra client config not found  - skip loading this plugin
+")
+
+cmd="examples/simple-agent/simple-agent --etcdv3-config=examples/etcdv3-lib/etcd.conf --kafka-config=examples/kafka-plugin/manual-partitioner/kafka.conf  --redis-config=examples/redis-lib/node-client.yaml --cassandra-config=examples/cassandra-lib/client-config.yaml"
+testOutput "${cmd}" "${expected}" "${unexpected}" 0 # the cmd continues to run - we will kill it later
+
+stopRedis >> /dev/null
+sleep 10
+docker exec etcd etcdctl get --prefix "" | grep  redis
+
+expected=("Agent plugin state update.*Get(/probe-redis-connection) failed: EOF.*status-check.*plugin=redis state=error
+")
+
+unexpected=("Agent plugin state update.*plugin=redis state=ok
+")
+
+testOutput "${cmd}" "${expected}" "${unexpected}" 0 # cmd unchanged - ASSERT disconnected
+
+startRedis >> /dev/null
+sleep 10
+docker exec etcd etcdctl get --prefix "" | grep redis
+
+expected=("Agent plugin state update.*plugin=redis state=ok
+")
+
+unexpected=("Agent plugin state update.*Get(/probe-redis-connection) failed: EOF.*status-check.*plugin=redis state=error 
+")
+
+testOutput "${cmd}" "${expected}" "${unexpected}" 0 # cmd unchanged - ASSERT connected AGAIN
+
+kill $CMD_PID > /dev/null
+rm ${TMP_FILE} > /dev/null
+rm ${TMP_FILE2} > /dev/null
+
+stopEtcd
+stopKafka
+stopRedis
+stopCassandra
 
 ##########################################################################
